improve documentation & readme

+++ src/lib.rs
@@ -12,11 +12,16 @@ pub mod ffi {
     include!("bindings.rs");
 }
 
+/// Used to represent one PCM sample in output data (conditional).
+/// 
+/// Normally a signed 16-bit integer (i16), but if the "float" feature is enabled,
+/// it's a 32-bit single precision float (f32).
 #[cfg(not(feature = "float"))]
 pub type Sample = i16;
 #[cfg(feature = "float")]
 pub type Sample = f32;
 
+/// Streaming iterator yielding frame data & references to decoded PCM samples.
 pub struct Decoder<'a> {
     data: &'a [u8],
     ffi_frame: ffi::mp3dec_frame_info_t,
@@ -24,6 +29,7 @@ pub struct Decoder<'a> {
     pcm: [Sample; ffi::MINIMP3_MAX_SAMPLES_PER_FRAME as usize],
 }
 
+/// Info about the current frame yielded by a [Decoder](struct.Decoder.html).
 pub struct Frame<'a> {
     /// Bitrate of this frame in kb/s.
     pub bitrate: u32,
@@ -32,7 +38,7 @@ pub struct Frame<'a> {
     pub channels: u32,
 
     /// MPEG layer of this frame.
-    pub mpeg_layer: i32, // TODO: Enumify
+    pub mpeg_layer: u32,
 
     /// Reference to the samples in this frame, copy if needed to allocate.
     pub samples: &'a [Sample],
@@ -65,6 +71,7 @@ impl<'a> Decoder<'a> {
     }
 
     /// Reads the next frame, if available.
+    /// If non-sample data (ex. ID3) is found it's skipped over until samples are found.
     pub fn next_frame(&mut self) -> Option<Frame> {
         unsafe {
             let mut samples =
@@ -79,7 +86,7 @@ impl<'a> Decoder<'a> {
                     channels: self.ffi_frame.channels as u32,
                     samples: self.pcm.get_unchecked(..samples as usize), // todo: feature?
                     sample_rate: self.ffi_frame.hz as u32,
-                    mpeg_layer: self.ffi_frame.layer,
+                    mpeg_layer: self.ffi_frame.layer as u32,
                     sample_count: samples,
                     source_len: self.ffi_frame.frame_bytes as usize,
                 })
@@ -95,15 +102,19 @@ impl<'a> Decoder<'a> {
     /// Useful when you want to, for example, calculate the audio length.
     ///
     /// It should be noted that the [samples](struct.Frame.html#structfield.sample_count)
-    /// in [Frame](struct.Frame.html) are an empty slice,
-    /// but you can still read its [sample_count](struct.Frame.html#structfield.sample_count).
+    /// in [Frame](struct.Frame.html) will be an empty slice since it's not decoding,
+    /// but you can still read its [sample_count](struct.Frame.html#structfield.sample_count),
+    /// which when zero will indicate that the current frame
+    /// does not contain any samples to be decoded.
+    /// Unlike [next_frame](struct.Frame.html#method.next_frame), it will **not** be skipped over
+    /// automatically, but you can still of course call `skip_frame(frame.source_len)` on it.
     pub fn peek_frame(&mut self) -> Option<Frame> {
         let samples = unsafe { self.ffi_decode_frame(ptr::null(), 0) as u32 };
         if self.ffi_frame.frame_bytes != 0 {
             Some(Frame {
                 bitrate: self.ffi_frame.bitrate_kbps as u32,
                 channels: self.ffi_frame.channels as u32,
-                mpeg_layer: self.ffi_frame.layer,
+                mpeg_layer: self.ffi_frame.layer as u32,
                 samples: &[],
                 sample_rate: self.ffi_frame.hz as u32,
                 sample_count: samples,
